backend/requirements.txt
# Dependências Python
flask
flask-cors
python-dotenv
openai
PyPDF2
gunicorn
